Python Program Base
Sprint 1, this file is a basic python program representing the logic behind my Vehicle Maintenance Log application. Interaction through calls in the console, this is the basic outline with the features for the application, and does not include GUI or Unit Testing. Includes sort methods and any data strucuture logic that may be needed(Priority Queue, Linked List with Node).