Final Application
Sprint 3, This is the final program/application. It has a GUI and includes Linked List Mimic, and Priority Queue Data structures. It also includes Insertion type Sort with the python method 'lambda', that sorts through the log, and appends the sorted entry to a new, sorted list/log. All errors or bugs have been fixed, and the program runs/operates as it should.