Final Unit Testing
Sprint 4, This is the Unit Testing for the Vehicle Maintenance Log Application. It includes Unit Testing for all the features of the program including, adding vehicles, checking vehicles, adding and viewing current and future maintenance logs, priority sort of future maintenance log, and insertion sorting of maintenance log. Because of the integration of the GUI directly into the python program file, the methods had to be edited to exclude the GUI use for Unit Testing.